Corrected a bug where colors were not passing trought the vertice but the information was correct in the shader pipeline.
Solution was to create the shader object which I forgot to add after load and compiling the PixelShader.hlsl

I also added a rotation on the triangle with matrice opperation on the constant buffer.

All of the above were added to the BeginFrame() method.

in the Initialize() function I added a blend state and a RasterizerState to not cull the triangle.