Added mouse and keyboard control
WASD to move the camera

Arrow keys to rotate the triangle

Right click + drag to orbit the camera around the triangle.

These operation were added in the handleMessage of Window.cpp. They also have their own function to calculate coord of the mouse and the window. Needed to calculate delta time.

Needed to update GraphicsEngine::BeginFrame to add window reference to get win height and width of the windows for the projection matrix

which I added getter and setter for the window in the window class.

which ended up needing to add the class declaration in the GraphicEngine.h